requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
black==23.7.0
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.6
commonmark==0.9.1
decorator==5.1.1
dictdiffer==0.8.1
dominate==2.5.2
elevenlabs==0.2.21
exceptiongroup==1.1.2
executing==1.2.0
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==2.10
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
mistletoe==0.7.2
multidict==6.0.4
mypy==1.4.1
mypy-extensions==1.0.0
notion-client==2.0.0
openai==0.27.8
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.9.1
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.12
pydub==0.25.1
Pygments==2.15.1
PyOgg==0.6.14a1
python-dotenv==1.0.0
python-slugify==4.0.1
python-telegram-bot==20.4
pytz==2023.3
requests==2.24.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
text-unidecode==1.3
tomli==2.0.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzlocal==2.1
urllib3==1.25.11
Wave==0.0.2
wcwidth==0.2.6
yarl==1.9.2
